Refactored inheritance
